Moved ProcId to codec package
